+++ src/tests/test_magic_cte.py
@@ -57,7 +57,7 @@ def test_infer_dependencies(ip, capsys):
     )
 
     assert result == expected
-    assert "Generating CTE with stored snippets : 'author_sub'" in out
+    assert "Generating CTE with stored snippets: 'author_sub'" in out
 
 
 TABLE_NAME_TYPO_ERR_MSG = """
@@ -127,7 +127,7 @@ def test_snippets_delete(ip, capsys):
     )
 
     out, _ = capsys.readouterr()
-    assert "Generating CTE with stored snippets : 'another_orders'" in out
+    assert "Generating CTE with stored snippets: 'another_orders'" in out
     result_del = ip.run_cell(
         "%sqlcmd snippets --delete-force-all another_orders"
     ).result
@@ -176,5 +176,5 @@ def test_query_syntax_error(ip):
             "SELECT last_name FROM author_sub;",
         )
 
-    assert excinfo.value.error_type == "UsageError"
+    assert excinfo.value.error_type == "RuntimeError"
     assert CTE_MSG.strip() in str(excinfo.value)